refactor: Update GitHub Actions workflows for improved compatibility and structure
- Change the build directory path separator to forward slash for consistency across environments.
- Switch the execution environment for code quality checks from Windows to Ubuntu for better compatibility with CI tools.
- Refactor paths in the code quality PR workflow to ensure correct lint report file access.
- Remove deprecated logger strategy files and update related includes to reflect the new structure.
