+++ .github/workflows/code-quality-pr.yml
@@ -17,11 +17,11 @@ permissions:
 env:
   FORCE_COLOR: 1  # Enable colored output in logs
   TERM: xterm-256color  # Terminal type for better formatting
-  CMAKE_BUILD_DIR: ${{ github.workspace }}\build
+  CMAKE_BUILD_DIR: ${{ github.workspace }}/build
 
 jobs:
   code-quality:
-    runs-on: windows-latest
+    runs-on: ubuntu-latest
     if: github.event.issue.pull_request && github.event.comment.body == 'run code quality'
     steps:
       - name: Get PR data
@@ -67,7 +67,7 @@ jobs:
         shell: bash
         continue-on-error: true
         run: |
-          REPORT_CONTENT=$(cat ${{ env.CMAKE_BUILD_DIR }}\lint-reports\github-report.txt)
+          REPORT_CONTENT=$(cat ${{ env.CMAKE_BUILD_DIR }}/lint-reports/github-report.txt)
           echo "report<<EOF" >> $GITHUB_OUTPUT
           echo "$REPORT_CONTENT" >> $GITHUB_OUTPUT
           echo "EOF" >> $GITHUB_OUTPUT
