+++ palantir-core/include/window/component/content_manager.hpp
@@ -123,14 +123,6 @@ public:
     }
 
     /**
-     * @brief Detect and notify observers of content size changes.
-     *
-     * This method should be called periodically to check content size and
-     * notify observers if the size has changed.
-     */
-    auto detectContentSizeChange() -> void override { pimpl_->detectContentSizeChange(); }
-
-    /**
      * @brief Get the current content width.
      *
      * @return int The current content width.
