+++ palantir-core/include/window/component/message/logger/log_message_mapper.hpp
@@ -9,7 +9,7 @@ namespace palantir::window::component::message::logger {
 
 class PALANTIR_CORE_API LogMessageMapper {
 public:
-    inline static auto fromJson(const nlohmann::json& json) -> LogMessageVO { return LogMessageVO{json.dump()}; }
+    static auto fromJson(const nlohmann::json& json) -> LogMessageVO { return LogMessageVO{json.dump()}; }
 };
 
 }  // namespace palantir::window::component::message::logger
